+++ src/main/scala/org/clulab/embeddings/word2vec/Word2vec.scala
@@ -9,7 +9,7 @@ import org.clulab.utils.MathUtils
 /**
  * Implements similarity metrics using the word2vec matrix
  * IMPORTANT: In our implementation, words are lower cased but NOT lemmatized or stemmed (see sanitizeWord)
- * User: mihais, dfried
+ * User: mihais, dfried, gus
  * Date: 11/25/13
  *
  */
@@ -23,8 +23,19 @@ class Word2Vec(matrixConstructor: => Map[String, Array[Double]]) {
   def this(mf: String, wordsToUse: Option[Set[String]] = None) = {
     this(Word2Vec.loadMatrix(mf, wordsToUse)._1)
   }
-  // construct the matrix for this instance lazily by calling the constructor
-  lazy val matrix : Map[String, Array[Double]] = matrixConstructor
+
+  /** alternate constructor to allow loading from a source, possibly with a set of words to constrain the vocab */
+  def this(src: io.Source, wordsToUse: Option[Set[String]]) = {
+    this(Word2Vec.loadMatrixFromSource(src, wordsToUse)._1)
+  }
+
+  /** alternate constructor to allow loading from a stream, possibly with a set of words to constrain the vocab */
+  def this(is: InputStream, wordsToUse: Option[Set[String]]) = {
+    this(Word2Vec.loadMatrixFromStream(is, wordsToUse)._1)
+  }
+
+  // laziness here causes problems with InputStream-based alternate constructor
+  val matrix : Map[String, Array[Double]] = matrixConstructor
 
   def saveMatrix(mf: String) {
     val pw = new PrintWriter(mf)
@@ -327,7 +338,7 @@ class Word2Vec(matrixConstructor: => Map[String, Array[Double]]) {
    */
   def sanitizedAvgSimilarity(t1:Iterable[String], t2:Iterable[String]):(Double, ArrayBuffer[(Double, String, String)]) = {
     // Top words
-    var pairs = new ArrayBuffer[(Double, String, String)]
+    val pairs = new ArrayBuffer[(Double, String, String)]
 
     var avg = 0.0
     var count = 0
@@ -444,20 +455,47 @@ object Word2Vec {
     }
   }
 
-  private def loadMatrix(mf:String, wordsToUse: Option[Set[String]]):(Map[String, Array[Double]], Int) = {
+  private def loadMatrix(mf: String, wordsToUse: Option[Set[String]]):(Map[String, Array[Double]], Int) = {
     logger.debug("Started to load word2vec matrix from file " + mf + "...")
+    val src: Source = Source.fromFile(mf, "iso-8859-1")
+    val lines: Iterator[String] = src.getLines()
+    val matrix = buildMatrix(lines, wordsToUse)
+    src.close()
+    logger.debug("Completed matrix loading.")
+    matrix
+  }
+
+  private def loadMatrixFromStream(is: InputStream, wordsToUse: Option[Set[String]]):(Map[String, Array[Double]], Int) = {
+    logger.debug("Started to load word2vec matrix from stream ...")
+    val src: Source = io.Source.fromInputStream(is, "iso-8859-1")
+    val lines: Iterator[String] = src.getLines
+    val matrix = buildMatrix(lines, wordsToUse)
+    src.close()
+    logger.debug("Completed matrix loading.")
+    matrix
+  }
+  private def loadMatrixFromSource(src: io.Source, wordsToUse: Option[Set[String]]):(Map[String, Array[Double]], Int) = {
+    logger.debug("Started to load word2vec matrix from source ...")
+    val lines: Iterator[String] = src.getLines()
+    val matrix = buildMatrix(lines, wordsToUse)
+    logger.debug("Completed matrix loading.")
+    matrix
+  }
+
+  private def buildMatrix(lines: Iterator[String], wordsToUse: Option[Set[String]]): (Map[String, Array[Double]], Int) = {
     val m = new collection.mutable.HashMap[String, Array[Double]]()
     var first = true
     var dims = 0
-    for((line, index) <- Source.fromFile(mf, "iso-8859-1").getLines().zipWithIndex) {
+
+    for((line, index) <- lines.zipWithIndex) {
       val bits = line.split("\\s+")
       if(first) {
         dims = bits(1).toInt
         first = false
       } else {
-          if (bits.length != dims + 1) {
-              println(s"${bits.length} != ${dims + 1} found on line ${index + 1}")
-              }
+        if (bits.length != dims + 1) {
+          println(s"${bits.length} != ${dims + 1} found on line ${index + 1}")
+        }
         assert(bits.length == dims + 1)
         val w = bits(0)
         if (wordsToUse.isEmpty || wordsToUse.get.contains(w)) {
@@ -475,7 +513,6 @@ object Word2Vec {
     logger.debug("Completed matrix loading.")
     (m.toMap, dims)
   }
-
 
   def main(args:Array[String]) {
     val w2v = new Word2Vec(args(0), None)